+++ docs/changes.rst
@@ -15,6 +15,9 @@ Version 0.5.3
 Unreleased.
 
  - Fixed alpha channel set to "on" & "off" values for ImageMagick-7. [:issue:`404`]
+ - Updated :meth:`Image.composite <wand.image.BaseImage.composite>` &
+   :meth:`Image.composite_channel <wand.image.BaseImage.composite_channel>` to
+   include optional arguments for composite methods that require extra controls.
  - Support for numpy arrays. [:issue:`65`]
      - Added :meth:`Image.from_array <wand.image.Image.from_array>` class method.
  - Support color map / palette manipulation. [:issue:`403`]